Combine storage classes from backends.storage.* with their respective backends in backends.*
