+++ requests_cache/backends/gridfs.py
@@ -1,11 +1,13 @@
-#!/usr/bin/env python
-"""
-    requests_cache.backends.gridfs
-    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+from gridfs import GridFS
+from pymongo import MongoClient
 
-    ``gridfs`` cache backend
+from .base import BaseCache, BaseStorage
+from .mongo import MongoDict
 
-    Use MongoDB GridFS to support documents greater than 16MB.
+
+class GridFSCache(BaseCache):
+    """GridFS cache backend.
+    Use this backend to store documents greater than 16MB.
 
     Usage:
         requests_cache.install_cache(backend='gridfs')
@@ -13,14 +15,7 @@
     Or:
         from pymongo import MongoClient
         requests_cache.install_cache(backend='gridfs', connection=MongoClient('another-host.local'))
-"""
-from .base import BaseCache
-from .storage.gridfspickledict import GridFSPickleDict
-from .storage.mongodict import MongoDict
-
-
-class GridFSCache(BaseCache):
-    """``gridfs`` cache backend."""
+    """
 
     def __init__(self, db_name, **options):
         """
@@ -30,3 +25,52 @@ class GridFSCache(BaseCache):
         super().__init__(**options)
         self.responses = GridFSPickleDict(db_name, options.get('connection'))
         self.redirects = MongoDict(db_name, 'redirects', self.responses.connection)
+
+
+class GridFSPickleDict(BaseStorage):
+    """A dictionary-like interface for a GridFS collection"""
+
+    def __init__(self, db_name, connection=None, **kwargs):
+        """
+        :param db_name: database name (be careful with production databases)
+        :param connection: ``pymongo.Connection`` instance. If it's ``None``
+                           (default) new connection with default options will
+                           be created
+        """
+        super().__init__(**kwargs)
+        if connection is not None:
+            self.connection = connection
+        else:
+            self.connection = MongoClient()
+
+        self.db = self.connection[db_name]
+        self.fs = GridFS(self.db)
+
+    def __getitem__(self, key):
+        result = self.fs.find_one({'_id': key})
+        if result is None:
+            raise KeyError
+        return self.deserialize(result.read())
+
+    def __setitem__(self, key, item):
+        self.__delitem__(key)
+        self.fs.put(self.serialize(item), **{'_id': key})
+
+    def __delitem__(self, key):
+        res = self.fs.find_one({'_id': key})
+        if res is not None:
+            self.fs.delete(res._id)
+
+    def __len__(self):
+        return self.db['fs.files'].count()
+
+    def __iter__(self):
+        for d in self.fs.find():
+            yield d._id
+
+    def clear(self):
+        self.db['fs.files'].drop()
+        self.db['fs.chunks'].drop()
+
+    def __str__(self):
+        return str(dict(self.items()))
